refactor: 사이드바 컨텐츠 분리
- 안건명 테이블 레이아웃 수정
- 계정 드롭다운 스타일 수정
